requirements/requirements_twc.txt
setuptools>=68
wheel
numpy==1.26.4
pandas==2.2.3
scipy>=1.10,<1.14
matplotlib>=3.1,<3.11
PyYAML>=6.0,<7
tqdm==4.67.3
tensorflow==2.17.0
